+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ master
 ~~~~~~
 
 1) [TASK] Simplify autoload.php with require instead of include.
+2) [TASK][BREAKING] Remove not needed public method requireFile.
 
 
 2.0.1